+++ src/providers/UserProvider.tsx
@@ -1,9 +1,11 @@
 import React, { createContext, FunctionComponent, useEffect, useState } from "react";
+import { RoleType } from "../api/RolesApi";
 import { IPerson, UserApiConfig } from "../api/UserApi";
 
 
 export interface IUserContext {
     user: IPerson | undefined,
+    roles: RoleType[],
     loadingUser: boolean
 }
 
@@ -12,6 +14,7 @@ export const UserContext = createContext<Partial<IUserContext>>({ user: undefine
 export const UserProvider: FunctionComponent = ({ children }) => {
     const [loading, setLoading] = useState(true);
     const [user, setUser] = useState<IPerson>();
+    const [roles, setRoles] = useState<RoleType[]>([]);
 
     const userApi = UserApiConfig.getApi();
 
@@ -19,6 +22,10 @@ export const UserProvider: FunctionComponent = ({ children }) => {
         const user = await userApi.getCurrentUser();
         if (user) {
             setUser(user);
+            let userRoles = await userApi.getCurrentUsersRoles();
+            if (userRoles) {
+                setRoles(userRoles);
+            }
         }
         setLoading(false);
     }
@@ -29,6 +36,7 @@ export const UserProvider: FunctionComponent = ({ children }) => {
 
     const userContext: IUserContext = {
         user,
+        roles,
         loadingUser: loading
     }